Refactored FlxNapeState into FlxNapeSpace
Previously it was impossible to use nape integration in conjunction with FlxUIStates, by refactoring FlxNapeState into a plugin this is now possible.
Instead of extending FlxNapeState, a FlxNapeSpace.init(); call is now required at the start of the state's create().

napeDebugEnabled has been renamed to drawDebug (consistent with FlxG.debugger.drawDebug).