Make TiledLayer strip leading whitespace from csvData
FlxTilemap blows up if the first character in a map CSV is a newline. On
at least the OS X 0.10.2 release of Tiled, layers are stored in .tmx
files with a leading newline before the actual CSV data. In an effort to
make FlxTilemap not blow up, let's just strip any leading whitespace
when we get csvData from TiledLayer.